backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
openai>=1.35.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
rapidfuzz>=3.9.0,<4.0.0
duckdb
pandas
openpyxl
numpy
langchain
langchain-openai
langchain-community
sqlalchemy
